feat(forms.py): remove posted on and parent field
The fields were removed for the following
reasons: 

The `posted on` is automatically set in the
model and was removed. The `parent` field 
will depend on which comment a user 
makes a reply to and the logic will be handled 
in the view- we cannot lets the users decide this.